requirements.txt
langchain
langchain_community
langchain_core
langchain-openai
python-dotenv
streamlit
chromadb
faiss-cpu
pypdf
python-docx
unstructured
